[CHORE](foundation-api): Split mullet_proxy tests into sibling
Move the `#[cfg(test)] mod tests` block out of `mullet_proxy.rs`
and into a sibling `mullet_proxy/tests.rs` file. The handler module
shrinks from ~560 lines (~70% tests) to ~150, keeping production
code easy to scan; tests stay co-located via Rust's submodule file
convention so they retain access to private items
(`merge_user`, `MulletProxyError`).
